Move glyph layout processing out of TTFFont and into it's own class (LayoutEngine)
TTFFont#glyphsForString is now just a simple glyph mapping (nothing contextual). A new method, TTFFont#layout, is used to do glyph mapping, substitution, and positioning all at once. A GlyphRun object is returned as a result, which includes the glyphs and positions.

Also adds support for selecting the script that is used during opentype shaping, and reverses the glyphs if the script is written right-to-left.