Re-encode variation glyphs in PDF subsets
PDF viewers don’t support TrueType variation glyphs (gvar, avar, etc.), so we need to re-encode the post-processed variants into the subset instead of the original outlines.

The glyph encoder is generic: it takes a Path object and outputs a Buffer containing a TrueType glyph. We can eventually use it when producing fonts from scratch, as opposed to just subsetting.