Proposal for #77 - support using the module in webpack
This uses the existing trie generation scripts to also generate a JS
module that exports the contents of the trie as a Buffer. This should
allow tools like webpack to successfully package fontkit and allow it to
be required as a module.

Prior to this, webpack could package it fine, but when the generated
bundle executes and tries to import fontkit, it throws an error because
of the load-time fs.readFileSync calls.